pyproject.toml
[project]
name = "pepperdine-gmail-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-api-python-client>=2.194.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "mcp[cli]>=1.27.0",
    "python-dotenv>=1.2.2",
]
